Move GoogleApi.ts to lib
