+++ lib/GoogleApi.ts
@@ -6,27 +6,27 @@ function getJwt() {
     return new google.auth.JWT(
         process.env.SERVICE_CLIENT_EMAIL, undefined, process.env.SERVICE_CLIENT_KEY,
         ['https://www.googleapis.com/auth/spreadsheets']
-    )
+    );
 }
 
 export type SheetContent =  GaxiosResponse<sheets_v4.Schema$ValueRange> | null | undefined
 
 export async function getSheet(spreadsheetId: string, range: string): Promise<SheetContent> {
-    const sheets = google.sheets({version: 'v4'})
+    const sheets = google.sheets({version: 'v4'});
     return new Promise((resolve, reject) => {
         sheets.spreadsheets.values.get({
             spreadsheetId,
             range,
             auth: getJwt(),
-            key: process.env.SHEETS_API_KEY
+            key: process.env.SHEETS_API_KEY,
         }, (err, res) => {
             if (err) {
-                reject(err)
+                reject(err);
             } else {
-                resolve(res)
+                resolve(res);
             }
-        })
-    })
+        });
+    });
 }
 
 export type SheetValues = GaxiosResponse<sheets_v4.Schema$ValueRange>['data']['values'];
@@ -37,7 +37,7 @@ export async function getSheetValues(spreadsheetId: string, range: string): Prom
 }
 
 export async function appendSheetValues(spreadsheetId: string, range: string, values: string[][], valueInputOption: 'RAW' | 'USER_ENTERED' = 'RAW') {
-    const sheets = google.sheets({version: 'v4'})
+    const sheets = google.sheets({version: 'v4'});
     return sheets.spreadsheets.values.append({
         spreadsheetId,
         range,
@@ -46,6 +46,6 @@ export async function appendSheetValues(spreadsheetId: string, range: string, va
         key: process.env.SHEETS_API_KEY,
         requestBody: {
             values,
-        }
-    })
+        },
+    });
 }
